switch to github releases
